# refactor(jobsdb): support executing commands in the context of an existing transaction