--- a/router/router.go
+++ b/router/router.go
@@ -151,7 +151,7 @@ type HandleT struct {
 	lastQueryRunTime time.Time
 	timeGained       float64
 
-	payloadLimit int64
+	payloadLimit     int64
 	transientSources transientsource.Service
 }
 
@@ -296,9 +296,6 @@ func (rt *HandleT) addResultSetMeta(resultSet *resultSetT) {
 			delete(rt.resultSetMeta, key)
 		}
 	}
-
-	//rt.logger.Infof("MEM LEAK DEBUG router %v Length of result set %v minResultSetId %v newResultSetID %v", rt.destName, len(rt.resultSetMeta), minResultSetID, newResultSet.id)
-
 }
 
 func (rt *HandleT) getResultSet(id int64) *resultSetT {
@@ -315,7 +312,6 @@ func isJobTerminated(status int) bool {
 	if status == 429 {
 		return false
 	}
-
 	return status >= 200 && status < 500
 }
 
@@ -1630,20 +1626,22 @@ func (rt *HandleT) commitStatusList(responseList *[]jobResponseT) {
 			rt.errorDB.Store(routerAbortedJobs)
 		}
 		//Update the status
-		txn := rt.jobsDB.BeginGlobalTransaction()
-		rt.jobsDB.AcquireUpdateJobStatusLocks()
-		err := rt.jobsDB.UpdateJobStatusInTxn(txn, statusList, []string{rt.destName}, nil)
+		err := rt.jobsDB.WithUpdateSafeTx(func(tx jobsdb.UpdateSafeTx) error {
+			err := rt.jobsDB.UpdateJobStatusInTx(tx, statusList, []string{rt.destName}, nil)
+			if err != nil {
+				rt.logger.Errorf("[Router] :: Error occurred while updating %s jobs statuses. Panicking. Err: %v", rt.destName, err)
+				return err
+			}
+			//Save msgids of aborted jobs
+			if len(jobRunIDAbortedEventsMap) > 0 {
+				GetFailedEventsManager().SaveFailedRecordIDs(jobRunIDAbortedEventsMap, tx.Tx())
+			}
+			rt.Reporting.Report(reportMetrics, tx.Tx())
+			return nil
+		})
 		if err != nil {
-			rt.logger.Errorf("[Router] :: Error occurred while updating %s jobs statuses. Panicking. Err: %v", rt.destName, err)
 			panic(err)
 		}
-		//Save msgids of aborted jobs
-		if len(jobRunIDAbortedEventsMap) > 0 {
-			GetFailedEventsManager().SaveFailedRecordIDs(jobRunIDAbortedEventsMap, txn)
-		}
-		rt.Reporting.Report(reportMetrics, txn)
-		rt.jobsDB.CommitTransaction(txn)
-		rt.jobsDB.ReleaseUpdateJobStatusLocks()
 	}
 
 	if rt.guaranteeUserEventOrder {
@@ -2070,8 +2068,7 @@ func destinationID(job *jobsdb.JobT) string {
 }
 
 func (*HandleT) crashRecover() {
-	//Perform any crash recover items here.
-	//None as of now
+	// NO-OP
 }
 
 func Init() {