# Description

Refactoring jobsdb in order to support executing commands (append jobs & update job statuses) in the context of an existing transaction.

In the context of this refactoring a number of tasks have been performed, most importantly:
- introduce `WithTx`, `WithStoreSafeTx`, `WithUpdateSafeTx` methods in JobsDB
- rename, reorganise and document JobsDB methods (see JobsDB interface)
- drop `globalDBHandle`
- drop unnecessary utility methods from JobsDB
- all `InTxn` commands now respect the writerQueue configuration
- use `SAVEPOINT` where appropriate
- setup connection pool limits based on jobsdb configuration
- unexport private-at-heart types
- update enterprise-commit
- use semaphore channels instead of goroutine workers for implementing reader/writer queues

## Review Guide

1. Please start by reading the new `jobsdb.JobsDB` interface definition
2. Read the rest of the pull request changes, top-to-bottom. As you read these changes you will find comments in key sections which there to guide you and explain critical changes. 

## Notion Ticket

[Refactor jobsdb operations to execute in the context of an existing transaction](https://www.notion.so/rudderstacks/Refactor-jobsdb-operations-to-execute-in-the-context-of-an-existing-transaction-227d67e9a8fd4e6db2ca26f71f6817f3)

[Refactor jobsdb reader/writer queue](https://www.notion.so/rudderstacks/Refactor-jobsdb-reader-writer-queue-3063893b909345f29b3165eb659a0b6f)

## Security

- [x] The code changed/added as part of this pull request won't create any security issues with how the software is being used.