@atzoum 2022-05-05
A performance benchmark (see `BenchmarkJobsdb` in `jobsdb/integration_test.go`) comparing the previous implementation with the new one indicates no regression.

```markdown
# Before
BenchmarkJobsdb/store_and_consume_10000_jobs_using_1_stream(s)-10         	       1	17571726041 ns/op	576593440 B/op	 6211230 allocs/op
BenchmarkJobsdb/store_and_consume_10000_jobs_using_10_stream(s)-10        	       1	6995922292 ns/op	579987200 B/op	 6259936 allocs/op
BenchmarkJobsdb/store_and_consume_10000_jobs_using_200_stream(s)-10       	       1	6704573667 ns/op	586918048 B/op	 6332132 allocs/op
BenchmarkJobsdb/store_and_consume_10000_jobs_using_500_stream(s)-10       	       1	7406804292 ns/op	593137704 B/op	 6406599 allocs/op

# After
BenchmarkJobsdb/store_and_consume_10000_jobs_using_1_stream(s)-10         	       1	17258266875 ns/op	576617376 B/op	 6213160 allocs/op
BenchmarkJobsdb/store_and_consume_10000_jobs_using_10_stream(s)-10        	       1	5147047875 ns/op	580197320 B/op	 6264283 allocs/op
BenchmarkJobsdb/store_and_consume_10000_jobs_using_200_stream(s)-10       	       1	5060879625 ns/op	583470256 B/op	 6294680 allocs/op
BenchmarkJobsdb/store_and_consume_10000_jobs_using_500_stream(s)-10       	       1	6563832125 ns/op	589489216 B/op	 6378574 allocs/op


```